90% only need to check kill + diff 5.6

+++ 1.Smash/signals.c
@@ -1,17 +1,24 @@
 // signals.c
 #include "signals.h"
 
-
+extern int signal_flag;
 
 //required functions for signal handling
 
 void ctrl_c_handler(int sig) {  //perhaps with arg int sig
     printf("smash: caught CTRL+C\n"); //perhaps after if condition
    // int fg_index = find_foreground_job();  // <-- function that returns index of foreground job, or -1 if none
-    if (fg_pid != getpid()) {
-        if (kill(-fg_pid, SIGKILL) == -1) {
-            perror("smash error: kill(SIGKILL)");// will not happen - if we see this print just delete the print line
-        }
+   if (fg_pid == getpid()){
+        signal_flag = 1;
+        //printf("smash > ");
+        return;
+    } 
+   
+   if (fg_pid != getpid()) {
+       if (kill(-fg_pid, SIGKILL) == -1)
+       {
+           perror("smash error: kill(SIGKILL)"); // will not happen - if we see this print just delete the print line
+       }
         else{
             printf("smash: process %d was killed\n", fg_pid); 
         }
@@ -21,7 +28,13 @@ void ctrl_c_handler(int sig) {  //perhaps with arg int sig
 void ctrl_z_handler(int sig) {  //perhaps with arg int sig
     printf("smash: caught CTRL+Z\n"); //perhaps after if condition
    // int fg_index = find_foreground_job();
-    if (fg_pid != getpid()) {
+   if (fg_pid == getpid()){
+    signal_flag = 1;
+    //printf("smash > ");
+    return;
+    } 
+   
+   if (fg_pid != getpid()) {
         if (kill(-fg_pid, SIGTSTP) == -1) {
             perror("smash error: kill(SIGTSTP)");
         }
